% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilyor.R
\name{OR_95CI}
\alias{OR_95CI}
\title{OR and 95 percent CI}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{coefficient}

\item{se}{standard error}

\item{siglevel}{significance level}

\item{roundto}{round number of digits of CI}
}
\value{
y
}
\description{
Creates odds ratio with 95 percent confidence intervals
}
\examples{
data.lrm <- lrm(y ~ x1 + x2, data = toydata, x = TRUE, y = TRUE)
OR_95CI(1.0055, 0.2534, 0.05, 2)
}
\author{
Lily Zhou
}
