% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
dataframe with 100obs and 3 variables
\describe{
   \item{y}{outcome}
  \item{x1}{continuous predictor variable}
  \item{x2}{binary predictor variable
  }
  }
}
\usage{
toydata
}
\description{
Toy data to be used for regression
}
\keyword{datasets}
